LayoutTests/platform/mac-catalina/imported/w3c/web-platform-tests/xhr/send-entity-body-get-head-expected.txt


PASS XMLHttpRequest: send() - non-empty data argument and GET/HEAD (GET)
FAIL XMLHttpRequest: send() - non-empty data argument and GET/HEAD (HEAD) assert_equals: expected "NO" but got "0"
